requirements.txt
aiogram>=3.4,<4.0
pillow>=10.0.0
python-dotenv>=1.0.0
